refactor(app): apply ts naming guideline to app routes and actions
