+++ app/src/hooks/treeconsole/actions/dialog.ts
@@ -9,7 +9,7 @@ import type { HierarchicalTreeNode } from '@hierarchidb/ui-treeconsole-base';
 import { DualKeyMap } from '@hierarchidb/util';
 import { loadUIPlugin } from '~/plugin-loaders/ui-plugin-loader';
 import type { TreeConsoleActionDeps } from '~/hooks/treeconsole/types';
-import { showCommandError } from './helpers.ts';
+import { showCommandError } from './treeConsoleActionUtils.ts';
 import { openInNewTab } from '~/utils/openInNewTab';
 
 export const PREVIEW_GUARD_NODE_TYPES = new Set([
